Move butter bar logger into core report
