+++ lighthouse-core/report/scripts/logger.js
